+++ services/doc.go
@@ -1,2 +1,0 @@
-// Package services helps to initialize services & applications.
-package services