+++ money/money.go
@@ -1,172 +1,173 @@
 package money
 
 import (
-	"math"
+	"fmt"
 	"strconv"
 	"strings"
 
-	"github.com/Rhymond/go-money"
 	"libs.altipla.consulting/errors"
 )
 
-func init() {
-	money.AddCurrency("EUR", "\u20ac", "1 $", ".", ",", 2)
+var (
+	EUR = FormatConfig{
+		Symbol:   "\u20ac",
+		Thousand: ",",
+	}
+
+	USD = FormatConfig{
+		Symbol:   "$",
+		Prefix:   true,
+		Thousand: ",",
+	}
+
+	GBP = FormatConfig{
+		Symbol:   "\u00a3",
+		Prefix:   true,
+		Thousand: ",",
+	}
+)
+
+type FormatConfig struct {
+	Symbol        string
+	Prefix        bool
+	Thousand      string
+	Decimal       string
+	ForceDecimals bool
 }
 
-// Money represents a monetary value
-type Money struct {
-	value *money.Money
+// Currency returns the configuration of a currency by its name. Unknown currencies will return the default config.
+func Currency(currency string) FormatConfig {
+	currencies := map[string]FormatConfig{
+		"EUR": EUR,
+		"USD": USD,
+		"GBP": GBP,
+	}
+
+	return currencies[currency]
 }
 
-// NewFromCents creates a new instance with a cents value.
-func NewFromCents(cents int64) *Money {
-	return &Money{money.New(cents, "")}
-}
+type Money int32
 
-// New creates a new instance with a zero value.
-func New() *Money {
-	return NewFromCents(0)
+// FromCents creates a new instance with a cents value.
+func FromCents(cents int32) Money {
+	return Money(cents)
 }
 
 // Parse a string to create a new money value. It can read `XX.YY` and `XX,YY`.
 // An empty string is parsed as zero.
-func Parse(s string) (*Money, error) {
+func Parse(s string) (Money, error) {
 	if len(s) == 0 {
-		return New(), nil
+		return Money(0), nil
 	}
 
 	s = strings.Replace(s, ",", ".", -1)
 
-	var amount int64
+	var amount int32
 	switch parts := strings.Split(s, "."); len(parts) {
 	case 1:
 		var err error
 		units, err := strconv.ParseInt(parts[0], 10, 64)
 		if err != nil {
-			return nil, errors.Trace(err)
+			return Money(0), errors.Trace(err)
 		}
-		amount = units * 100
+		amount = int32(units) * 100
 
 	case 2:
 		units, err := strconv.ParseInt(parts[0], 10, 64)
 		if err != nil {
-			return nil, errors.Trace(err)
+			return Money(0), errors.Trace(err)
 		}
 		if len(parts[1]) > 2 {
 			parts[1] = parts[1][:2]
 		}
 		decimals, err := strconv.ParseInt(parts[1], 10, 64)
 		if err != nil {
-			return nil, errors.Trace(err)
+			return Money(0), errors.Trace(err)
 		}
 		if len(parts[1]) == 1 {
-			amount = units*100 + decimals*10
+			amount = int32(units)*100 + int32(decimals)*10
 		} else {
-			amount = units*100 + decimals
+			amount = int32(units)*100 + int32(decimals)
 		}
 
 	default:
-		return nil, errors.Errorf("cannot parse money value: %v", s)
+		return Money(0), errors.Errorf("cannot parse money value: %v", s)
 	}
 
-	return NewFromCents(amount), nil
+	return Money(amount), nil
 }
 
-// Cents returns the value with cents precision (2 decimal places) as a number.
-func (m *Money) Cents() int64 {
-	return m.value.Amount()
-}
+// Format the money value with a specific configuration.
+func (m Money) Format(config FormatConfig) string {
+	if config.Decimal == "" {
+		config.Decimal = "."
+	}
 
-func (m *Money) formatVal(prec int) int64 {
-	value := m.value.Amount()
-	switch {
-	case prec > 2:
-		value *= int64(math.Pow10(prec - 2))
-
-	case prec == 1:
-		if value%10 > 5 {
-			value = (value / 10) + 1
-		} else {
-			value = value / 10
+	value := strconv.FormatInt(int64(m), 10)
+	value = fmt.Sprintf("%03v", value)
+	if config.Thousand != "" {
+		for i := len(value) - 5; i > 0; i -= 3 {
+			value = value[:i] + config.Thousand + value[i:]
 		}
-
-	case prec == 0:
-		value = value / 100
 	}
 
+	value = value[:len(value)-2] + config.Decimal + value[len(value)-2:]
+
+	if config.Symbol != "" {
+		if config.Prefix {
+			value = config.Symbol + value
+		} else {
+			value = value + " " + config.Symbol
+		}
+	}
+
+	if !config.ForceDecimals {
+		value = strings.TrimSuffix(value, config.Decimal+"00")
+	}
 	return value
 }
 
-// Format the money value with a specific decimal precision.
-func (m *Money) Format(prec int) string {
-	cur := money.GetCurrency("EUR")
-	f := money.NewFormatter(prec, cur.Decimal, "", cur.Grapheme, "1")
-	return f.Format(m.formatVal(prec))
+// Format the money value with the default configuration
+func (m Money) String() string {
+	return m.Format(FormatConfig{})
 }
 
-// Format the money value with a specific decimal precision and thousands separator.
-func (m *Money) FormatHuman(prec int) string {
-	cur := money.GetCurrency("EUR")
-	f := money.NewFormatter(prec, cur.Decimal, cur.Thousand, cur.Grapheme, "1")
-	return f.Format(m.formatVal(prec))
-}
-
-// Mul multiplies the money value n times and returns the result.
-func (m *Money) Mul(n int64) *Money {
-	return &Money{m.value.Multiply(n)}
-}
-
-// Add two money values together and returns the result.
-func (m *Money) Add(other *Money) *Money {
-	result, err := m.value.Add(other.value)
-	if err != nil {
-		panic(err)
-	}
-
-	return &Money{result}
-}
-
-// Sub subtracts two money values and returns the result.
-func (m *Money) Sub(other *Money) *Money {
-	result, err := m.value.Subtract(other.value)
-	if err != nil {
-		panic(err)
-	}
-
-	return &Money{result}
-}
-
-// Div divides two money values and returns the result.
-func (m *Money) Div(other *Money) *Money {
-	return &Money{m.value.Divide(other.value.Amount())}
-}
-
-// LessThan returns true if a money value is less than the other.
-func (m *Money) LessThan(other *Money) bool {
-	result, err := m.value.LessThan(other.value)
-	if err != nil {
-		panic(err)
-	}
-
-	return result
-}
-
-// AddTaxPercent adds a percentage of the price to itself.
-func (m *Money) AddTaxPercent(tax int64) *Money {
-	result, err := m.value.Multiply(tax).Divide(100).Add(m.value)
-	if err != nil {
-		panic(err)
-	}
-
-	return &Money{result}
+// Cents returns the value with cents precision (2 decimal places) as a number.
+func (m Money) Cents() int32 {
+	return int32(m)
 }
 
 // IsZero returns true if there is no money.
-func (m *Money) IsZero() bool {
-	return m.value.IsZero()
+func (m Money) IsZero() bool {
+	return m == 0
 }
 
-// Display formats the money value with a specific currency.
-func (m *Money) Display(currency string) string {
-	return money.GetCurrency(currency).Formatter().Format(m.value.Amount())
+// LessThan returns true if a money value is less than the other.
+func (m Money) LessThan(other Money) bool {
+	return m < other
+}
+
+// Mul multiplies the money value n times and returns the result.
+func (m Money) Mul(n int32) Money {
+	return Money(int32(m) * n)
+}
+
+// Add two money values together and returns the result.
+func (m Money) Add(other Money) Money {
+	return Money(int32(m) + int32(other))
+}
+
+// Sub subtracts two money values and returns the result.
+func (m Money) Sub(other Money) Money {
+	return Money(int32(m) - int32(other))
+}
+
+// Div divides two money values and returns the result.
+func (m Money) Div(other Money) Money {
+	return Money(int32(m) / int32(other))
+}
+
+// AddTaxPercent adds a percentage of the price to itself.
+func (m Money) AddTaxPercent(tax int32) Money {
+	return m.Mul(tax).Div(Money(100)).Add(m)
 }