+++ money/money_test.go
@@ -7,52 +7,46 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-func TestCents(t *testing.T) {
-	money := NewFromCents(12579)
-
-	require.EqualValues(t, money.Cents(), 12579)
-}
-
 func TestParse(t *testing.T) {
 	money, err := Parse("125,79")
 	require.NoError(t, err)
 
-	require.EqualValues(t, money.Cents(), 12579)
+	require.EqualValues(t, money, 12579)
 }
 
 func TestParseMultipleDecimals(t *testing.T) {
 	money, err := Parse("125.7923")
 	require.NoError(t, err)
 
-	require.EqualValues(t, money.Cents(), 12579)
+	require.EqualValues(t, money, 12579)
 }
 
 func TestParseMultipleDecimalsNoRound(t *testing.T) {
 	money, err := Parse("125.7963")
 	require.NoError(t, err)
 
-	require.EqualValues(t, money.Cents(), 12579)
+	require.EqualValues(t, money, 12579)
 }
 
 func TestParseFloatError(t *testing.T) {
 	money, err := Parse("10.03")
 	require.NoError(t, err)
 
-	require.EqualValues(t, money.Cents(), 1003)
+	require.EqualValues(t, money, 1003)
 }
 
 func TestParseOneDecimal(t *testing.T) {
 	money, err := Parse("10.3")
 	require.NoError(t, err)
 
-	require.EqualValues(t, money.Cents(), 1030)
+	require.EqualValues(t, money, 1030)
 }
 
 func TestParseOneDecimalWithZero(t *testing.T) {
 	money, err := Parse("10.30")
 	require.NoError(t, err)
 
-	require.EqualValues(t, money.Cents(), 1030)
+	require.EqualValues(t, money, 1030)
 }
 
 func TestParseFloatErrorIterative(t *testing.T) {
@@ -65,7 +59,7 @@ func TestParseFloatErrorIterative(t *testing.T) {
 		money, err := Parse(s)
 		require.NoError(t, err)
 
-		require.EqualValues(t, money.Cents(), i, "i: %v; s: %v", i, s)
+		require.EqualValues(t, money, i, "i: %v; s: %v", i, s)
 	}
 }
 
@@ -73,105 +67,136 @@ func TestParseWithoutDecimals(t *testing.T) {
 	money, err := Parse("10")
 	require.NoError(t, err)
 
-	require.EqualValues(t, money.Cents(), 1000)
+	require.EqualValues(t, money, 1000)
 }
 
-func TestFormatPrecisionFour(t *testing.T) {
-	money := NewFromCents(12345)
+func TestFormatDefault(t *testing.T) {
+	money := Money(12345)
 
-	require.EqualValues(t, money.Format(4), "123.4500")
+	require.EqualValues(t, money.Format(FormatConfig{}), "123.45")
 }
 
-func TestFormatPrecisionThree(t *testing.T) {
-	money := NewFromCents(12345)
+func TestFormatSymbol(t *testing.T) {
+	money := Money(12345)
 
-	require.EqualValues(t, money.Format(3), "123.450")
+	require.EqualValues(t, money.Format(EUR), "123.45 €")
 }
 
-func TestFormatPrecisionTwo(t *testing.T) {
-	money := NewFromCents(12345)
+func TestFormatSymbolPrefix(t *testing.T) {
+	money := Money(12345)
 
-	require.EqualValues(t, money.Format(2), "123.45")
+	require.EqualValues(t, money.Format(USD), "$123.45")
 }
 
-func TestFormatPrecisionOne(t *testing.T) {
-	money := NewFromCents(12346)
+func TestFormatNegativeSymbolPrefix(t *testing.T) {
+	money := Money(-12345)
 
-	require.EqualValues(t, money.Format(1), "123.5")
+	cnf := FormatConfig{
+		Symbol: "$",
+		Prefix: true,
+	}
+	require.EqualValues(t, money.Format(cnf), "$-123.45")
 }
 
-func TestFormatPrecisionZero(t *testing.T) {
-	money := NewFromCents(12346)
+func TestFormatThousand(t *testing.T) {
+	money := Money(123456)
 
-	require.EqualValues(t, money.Format(0), "123")
+	require.EqualValues(t, money.Format(EUR), "1,234.56 €")
 }
 
-func TestFormatWithoutThousandsSeparator(t *testing.T) {
-	money := NewFromCents(106400)
+func TestFormatCents(t *testing.T) {
+	money := Money(12)
 
-	require.EqualValues(t, money.Format(2), "1064.00")
+	require.EqualValues(t, money.Format(FormatConfig{}), "0.12")
 }
 
-func TestFormatHuman(t *testing.T) {
-	money := NewFromCents(106400)
+func TestFormatZeroDecimals(t *testing.T) {
+	money := Money(100)
 
-	require.EqualValues(t, money.FormatHuman(2), "1,064.00")
+	require.EqualValues(t, money.Format(FormatConfig{}), "1")
 }
 
-func TestMul(t *testing.T) {
-	money := NewFromCents(1000)
+func TestFormatForceDecimals(t *testing.T) {
+	money := Money(100)
 
-	require.EqualValues(t, money.Mul(3).Cents(), 3000)
+	cnf := FormatConfig{
+		ForceDecimals: true,
+	}
+	require.EqualValues(t, money.Format(cnf), "1.00")
 }
 
-func TestAdd(t *testing.T) {
-	money := NewFromCents(10000)
+func TestFormatCurrencyEUR(t *testing.T) {
+	money := Money(123456)
 
-	require.EqualValues(t, money.Add(NewFromCents(5000)).Cents(), 15000)
+	require.EqualValues(t, money.Format(Currency("EUR")), "1,234.56 €")
 }
 
-func TestSub(t *testing.T) {
-	money := NewFromCents(10000)
+func TestFormatCurrencyUSD(t *testing.T) {
+	money := Money(123456)
 
-	require.EqualValues(t, money.Sub(NewFromCents(4000)).Cents(), 6000)
+	require.EqualValues(t, money.Format(Currency("USD")), "$1,234.56")
 }
 
-func TestDiv(t *testing.T) {
-	money := NewFromCents(1234)
+func TestFormatCurrencyGBP(t *testing.T) {
+	money := Money(123456)
 
-	require.EqualValues(t, money.Div(NewFromCents(2)).Cents(), int64(617))
+	require.EqualValues(t, money.Format(Currency("GBP")), "£1,234.56")
+}
+
+func TestFormatEmptyCurrency(t *testing.T) {
+	money := Money(123456)
+
+	require.EqualValues(t, money.Format(Currency("")), "1234.56")
+}
+
+func TestFormatUnknownCurrency(t *testing.T) {
+	money := Money(123456)
+
+	require.EqualValues(t, money.Format(Currency("PLN")), "1234.56")
 }
 
 func TestLessThan(t *testing.T) {
-	money := NewFromCents(10000)
-	other := NewFromCents(5000)
+	money := Money(10000)
+	other := Money(5000)
 
 	require.False(t, money.LessThan(other))
 	require.True(t, other.LessThan(money))
 }
 
-func TestAddTaxPercent(t *testing.T) {
-	money := NewFromCents(10000)
-
-	require.EqualValues(t, money.AddTaxPercent(20).Cents(), 12000)
-}
-
 func TestIsZero(t *testing.T) {
-	money := New()
+	money := Money(0)
 	require.True(t, money.IsZero())
 
-	money = money.Add(NewFromCents(1))
+	money = Money(1)
 	require.False(t, money.IsZero())
 }
 
-func TestDisplayEUR(t *testing.T) {
-	money := NewFromCents(12345)
+func TestMul(t *testing.T) {
+	money := Money(1000)
 
-	require.Equal(t, money.Display("EUR"), "123.45 €")
+	require.EqualValues(t, money.Mul(3), 3000)
 }
 
-func TestDisplayUSD(t *testing.T) {
-	money := NewFromCents(12345)
+func TestAdd(t *testing.T) {
+	money := Money(10000)
 
-	require.Equal(t, money.Display("USD"), "$123.45")
+	require.EqualValues(t, money.Add(Money(5000)), 15000)
+}
+
+func TestSub(t *testing.T) {
+	money := Money(10000)
+
+	require.EqualValues(t, money.Sub(Money(4000)), 6000)
+}
+
+func TestDiv(t *testing.T) {
+	money := Money(1234)
+
+	require.EqualValues(t, money.Div(Money(2)), int32(617))
+}
+
+func TestAddTaxPercent(t *testing.T) {
+	money := Money(10000)
+
+	require.EqualValues(t, money.AddTaxPercent(20), 12000)
 }