+++ templates/funcs/money.go
@@ -3,9 +3,9 @@ package funcs
 import "libs.altipla.consulting/money"
 
 func Price(value int32) string {
-	return money.NewFromCents(int64(value)).Format(2)
+	return money.FromCents(value).Format(money.FormatConfig{})
 }
 
 func Money(currency string, value int32) string {
-	return money.NewFromCents(int64(value)).Display(currency)
+	return money.FromCents(value).Format(money.Currency(currency))
 }