+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This changelog will list only breaking changes in our APIs.
 ## v1.54.0
 
 - **connect:** Deleted without possible replacement. It is recommended to copy it externally.
+- **services:** Deleted without possible replacement. It is recommended to copy it externally.
 
 
 ## v1.52.0